enable/improve material sort and implement vertex index sorting for standard table elements
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@1865 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ pin/hitable.h
@@ -15,5 +15,5 @@ public:
 
     virtual bool IsTransparent()                        { return false; }
     virtual float GetDepth(const Vertex3Ds& viewDir)    { return 0.0f; }
-	virtual size_t GetMaterialID()                      { return 0; }
+	virtual unsigned long long GetMaterialID()          { return 0; }
 	};
